+++ templates/examples/extensions/index.js
@@ -172,6 +172,7 @@ module.exports=Object.assign(
         }],
         "ManagedPolicyArns": [
             "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole",
+            "arn:aws:iam::aws:policy/TranslateReadOnly",
             "arn:aws:iam::aws:policy/service-role/AWSLambdaVPCAccessExecutionRole",
             "arn:aws:iam::aws:policy/AWSXRayDaemonWriteAccess",
             "arn:aws:iam::aws:policy/AmazonKendraReadOnlyAccess"]
@@ -215,7 +216,7 @@ function jslambda(name){
           "Fn::If": [ "XRAYEnabled", {"Mode": "Active"},
               {"Ref" : "AWS::NoValue"} ]
       },
-      "Tags":[{
+            "Tags":[{
           Key:"Type",
           Value:"LambdaHook"
       }]
@@ -251,15 +252,15 @@ function pylambda(name){
       "Runtime": "python3.6",
       "Timeout": 300,
       "VpcConfig" : {
-          "Fn::If": [ "VPCEnabled", {
-              "SubnetIds": { "Fn::Split" : [ ",", {"Ref": "VPCSubnetIdList"} ] },
-              "SecurityGroupIds": { "Fn::Split" : [ ",", {"Ref": "VPCSecurityGroupIdList"} ] },
-          }, {"Ref" : "AWS::NoValue"} ]
-      },
-      "TracingConfig" : {
-          "Fn::If": [ "XRAYEnabled", {"Mode": "Active"},
-              {"Ref" : "AWS::NoValue"} ]
-      },
+        "Fn::If": [ "VPCEnabled", {
+            "SubnetIds": { "Fn::Split" : [ ",", {"Ref": "VPCSubnetIdList"} ] },
+            "SecurityGroupIds": { "Fn::Split" : [ ",", {"Ref": "VPCSecurityGroupIdList"} ] },
+        }, {"Ref" : "AWS::NoValue"} ]
+    },
+    "TracingConfig" : {
+        "Fn::If": [ "XRAYEnabled", {"Mode": "Active"},
+            {"Ref" : "AWS::NoValue"} ]
+    },
       "Tags":[{
           Key:"Type",
           Value:"LambdaHook"