+++ templates/examples/extensions/js_lambda_hooks/KendraFallback/translate.js
@@ -74,12 +74,11 @@ exports.translate_hit = async function(hit,usrLang,req){
     console.log("translate_hit:", JSON.stringify(hit,null,2));
     let hit_out = _.cloneDeep(hit);
     let a = _.get(hit, "a");
-    let markdown = _.get(hit, "alt.markdown");
-    let ssml = _.get(hit, "alt.ssml");
-    let rp = _.get(hit, "rp");
+    let markdown = _.get(hit, "markdown");
+    let ssml = _.get(hit, "ssml");
     let r = _.get(hit, "r");
     // catch and log errors before throwing exception.
-    if (a && _.get(hit,'autotranslate.a')) {
+    if (a) {
         try {
             hit_out.a = await get_translation(hit_out.a, usrLang,req);
         } catch (e) {
@@ -87,64 +86,24 @@ exports.translate_hit = async function(hit,usrLang,req){
             throw (e);
         }
     }
-    if (markdown && _.get(hit,'autotranslate.alt.markdown')) {
+    if (markdown) {
         try {
-            const res = await get_translation(hit_out.alt.markdown, usrLang,req);
-            hit_out.alt.markdown  = replaceAll(res,'] (http', '](http');
+            const res = await get_translation(hit_out.markdown, usrLang,req);
+            hit_out.markdown  = replaceAll(res,'] (http', '](http');
         } catch (e) {
             console.log("ERROR: Markdown caused Translate exception: ", a)
             throw (e);
         }
     }
-    if (ssml && _.get(hit,'autotranslate.alt.ssml')) {
+    if (ssml) {
         try {
-            hit_out.alt.ssml = await get_translation(hit_out.alt.ssml, usrLang,req);
+            hit_out.ssml = await get_translation(hit_out.ssml, usrLang,req);
         } catch (e) {
             console.log("ERROR: SSML caused Translate exception: ", a);
             throw (e);
         }
     }
-    if (rp && _.get(hit,'autotranslate.rp')) {
-        try {
-            hit_out.rp = await get_translation(hit_out.rp, usrLang);
-        } catch (e) {
-            console.log("ERROR: Reprompt caused Translate exception: ", rp);
-            throw (e);
-        }
-    }
-    if (r) {
-        try {
-            if (r.subTitle && r.subTitle.length > 0  && _.get(hit,'autotranslate.r.subTitle')) {
-                hit_out.r.subTitle = await get_translation(hit_out.r.subTitle, usrLang,req) ;
-            }
-            if (r.title && r.title.length > 0 && _.get(hit,'autotranslate.r.title')) {
-                hit_out.r.title = await get_translation(hit_out.r.title, usrLang,req) ;
-            }
-            if (r.text && r.text.length > 0) {
-                // no op
-            }
-            if (r.imageUrl && r.imageUrl.length > 0) {
-                // no op
-            }
-            if (r.url && r.url.length > 0) {
-                // no op
-            }
-            if (r.buttons && r.buttons.length > 0) {
-                for (let x=0; x<r.buttons.length; x++) {
-                    if (_.get(hit,'autotranslate.r.buttons[x].text')) {
-                        hit_out.r.buttons[x].text = await get_translation(hit_out.r.buttons[x].text, usrLang,req) ;
-                    }
-                    if (_.get(hit,'autotranslate.r.buttons[x].value')) {
-                        hit_out.r.buttons[x].value = await get_translation(hit_out.r.buttons[x].value, usrLang,req) ;
-                    }                    
-                }
-            }
-        } catch (e) {
-            console.log("ERROR: response card fields format caused Translate exception. Check syntax: " + e );
-            throw (e);
-        }
 
-    }
     console.log("Preprocessed Result: ", hit_out);
     return hit_out;    
 };